Tweaks to Ranks code
The Ranks code is more specific in references now. The reason being is I feel (for now) the best way to join ranks filtered by season is to store each via WITH and then join them together later. For example after the WITH clauses, this is run to show everything:
```
SELECT
TOS.*,
TOS_S01.*,
TOS_S02.*,
TOS_S03.*
FROM TOS
	LEFT JOIN TOS_S01 ON TOS_S01.Score_01 = TOS.Score
	LEFT JOIN TOS_S02 ON TOS_S02.Score_02 = TOS.Score
	LEFT JOIN TOS_S03 ON TOS_S03.Score_03 = TOS.Score
```
One issue at the moment is the single-season counts and averages are lost this way, but I am confident a solution can be found.