Added R-Shiny Visualizer
I have built up a visualizer in R-Shiny for showing stats, groups, and graphs of the data. It does require R, which is free and open source as well as found at R-Project.org. When first running the script, it should (ask to) install the necessary packages for it to run.