Improved "Episode" column support
The R-Shiny applet now checks for columns with "Episode" in the column name, instead of watching specifically for "Episode Production" and "Episode Air" (or "Episode.Production" and "Episode.Air" as R replaces spaces with dots in column names). This way additional episode orders and orders with other names can be recorded in the tables, though they must have "Episode" in the name, and preferably by "Episode " with the space (or "Episode." with the dot), as this will be removed from the name in the UI.
Also the Plots graph has been updated to rename "S01", "S02", etc. to be "Season 01" and "Season 02", etc.