Fixed Height for Brush tables and Better Season Strings
The tables that are shown based on the click brushes for the Plots and Histograms are now in a box of fixed height. This way the page height should remain constant, whether the tables are there or not.
The season information in the Histogram brush table is also now using the str_SEASON function. This way it will show "Season 1" instead of "Season 01" (assuming the season code was written "S01"), which is cleaner and consistent with the rest of the applet.