Toggle for Host facets on Plots
The Plots tab now has a "Separate Hosts" toggle. This allows the faceting by Host to be turned On and Off (On by default). When On, each host or rating has a row of plots showing their scores. When Off, the scores are not separated by source, though are still color coded. (There is no legend for what the colors are, but they are the same as when Host facets are enabled.)